+++ soccer/src/soccer/strategy/agent/position/marker.cpp
@@ -25,7 +25,7 @@ void Marker::choose_target(const WorldState* ws) {
     // If we ever use multiple Markers, they should choose different
     // robots to track from each other. Logic for this operation must be
     // added because multiple markers currently mark the same robot.
-    for (int i = 0; i < kNumShells; i++) {
+    for (int i = 0; i < (int) kNumShells; i++) {
         if (std::fabs(ws->get_robot(false, i).pose.position().x()) < marker_follow_cutoff &&
             ws->get_robot(false, i).pose.position().y() < y_bound &&
             (ws->ball.position - ws->get_robot(false, i).pose.position()).mag() > .25) {
